rename state to graph
