+++ packages/server/src/checkin/graph.rs
